fix NPE when an Activity is launched twice by multithread
When an activity is launching, it will go through preExecute -> execute -> postExecute. If there are two different threads launching the same Activity, the first thread do preExecute and addLaunchingActivity(token,r) into mLaunchingActivities, then the second thread also do preExecute and add the same token again. Then the first thread do execute and postExecute and removeLaunchingActivity(token) from mLaunchingActivities. After that, the second thread do execute to getLaunchingActivity(token) and get a null to pass into the handleLaunchingActivity, which causes NPE.

Therefore, to avoid the NPE when launching an activity in the case of multithread, we should identify whether the ActivityClientRecord get from the getLaunchingActivity(token) in execute is null. If it is null, then not go into the handleLaunchingActivity.

Change-Id: Id3b23193edf4066fab7000771bfb4d132f57d5d1
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>