Reduce the number of binders
In the APerformanceHintManager.createSession method, the binder will be sent to the peer HintManagerService.BinderService.createSession. In the HintManagerService.BinderService.createSession, it will be judged whether |mPreferredRateNanos| is equal to -1. If it is equal to -1, null will be returned. In order to prevent binder waste, Early judgment in APerformanceHintManager.createSession function

Signed-off-by: Liangliang Sui <coolsui.coding@gmail.com>
Change-Id: I0d97be7b15e08ced6e4c3f3c6b4c50e2104ecebe