Switch back to the application on /system when UID and path changed
If the application is a system update app, and the UID of application on /system is changed by OTA, and the version on /system is smaller than the version on /data.
It will remove disabledPkgSetting, and install as a normal app, so need swich back to the application on /system.

Bug: 260683094
Test: mamual
Change-Id: I90431cac15c8dde66eee81c42c91841bd92a96cf
Signed-off-by: Jason Edson <jaysonedson@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>